requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.5
django-allauth==65.4.1
django-ckeditor==6.7.2
django-environ==0.12.0
django-heroku==0.3.1
django-js-asset==3.0.1
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
